+++ packages/nesd/test/ui/emulator/nes_controller_swap_duty_cycles_test.dart
@@ -55,6 +55,7 @@ void main() {
     when(() => settings.volume).thenReturn(1.0);
     when(() => settings.lowPassFilter).thenReturn(false);
     when(() => settings.swapDutyCycles).thenReturn(swapDutyCycles);
+    when(() => settings.oamCorruption).thenReturn(true);
     when(() => settings.mixer).thenReturn(const MixerSettings());
     when(() => settings.fastForwardSpeed).thenReturn(FastForwardSpeed.x2);
     when(() => settings.autoSave).thenReturn(false);