+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Color Dreams (mapper 11) support
 - The web version can be installed to the home screen with offline support
 - Screenshot action (F12 by default) that saves the current frame 
+- OAM corruption: sprites glitch like on real hardware, can be turned off under Settings > General > Emulation
 
 ### Fixed
 - Fixed minor bugs in multiple CPU instructions